requirements-test.txt
# for main app
osbot-fast-api-serverless


# for pytest
pytest
pytest-cov


# for fastapi testing
httpx
requests
